+++ R/ColumnLinkedMatrix.R
@@ -30,7 +30,9 @@ subset.ColumnLinkedMatrix <- function(x, i, j, ..., drop) {
     }
     originalOrder <- (1:p)[order(j)]
     sortedColumns <- sort(j)
-    Z <- matrix(nrow = n, ncol = p, NA)
+    # Initialize result matrix as integer matrix because it does not take up as
+    # much space as double() but is more useful than logical()
+    Z <- matrix(data = integer(), nrow = n, ncol = p)
     # Use dimnames instead of rownames and colnames to avoid copy
     dimnames(Z) <- list(rownames(x)[i], colnames(x)[j])
     INDEX <- index(x)[sortedColumns, , drop = FALSE]