+++ R/ColumnLinkedMatrix.R
@@ -28,13 +28,21 @@ subset.ColumnLinkedMatrix <- function(x, i, j, ..., drop) {
     if (p > pX | n > nX) {
         stop("Either the number of columns or number of rows requested exceed the number of rows or columns in x, try dim(x)...")
     }
-    originalOrder <- (1:p)[order(j)]
-    sortedColumns <- sort(j)
     # Initialize result matrix as integer matrix because it does not take up as
     # much space as double() but is more useful than logical()
     Z <- matrix(data = integer(), nrow = n, ncol = p)
     # Use dimnames instead of rownames and colnames to avoid copy
     dimnames(Z) <- list(rownames(x)[i], colnames(x)[j])
+    # Providing a sorted column index will eliminate the need to reorder the
+    # result matrix later (avoiding a copy)
+    isUnsorted <- is.unsorted(i)
+    if (isUnsorted) {
+        # Reorder columns for sequential retrieval by chunk
+        originalOrder <- (1:p)[order(j)]
+        sortedColumns <- sort(j)
+    } else {
+        sortedColumns <- j
+    }
     INDEX <- index(x)[sortedColumns, , drop = FALSE]
     whatChunks <- unique(INDEX[, 1])
     end <- 0
@@ -45,7 +53,7 @@ subset.ColumnLinkedMatrix <- function(x, i, j, ..., drop) {
         # Convert to matrix to support data frames
         Z[, ini:end] <- as.matrix(x[[k]][i, TMP[, 3], drop = FALSE])
     }
-    if (length(originalOrder) > 1) {
+    if (isUnsorted) {
         Z[] <- Z[, originalOrder]
     }
     if (drop == TRUE && (n == 1 || p == 1)) {