+++ NEWS.md
@@ -3,6 +3,8 @@
 * Optimize subsetting by avoiding copies:
   * the return matrix is initialized as `integer()` instead of `logical()`
   * if row or column indices are sorted, do not resort return matrix
+  * if data from a single node is requested, pass through result instead of
+    buffering it in return matrix
 
 # LinkedMatrix 1.1.0